LPS-61952 Move com.liferay.portal.verify.model to com.liferay.portal.kernel.verify.model
